requirements.txt
twisted
mysql-connector-python